feat #56: 그룹 가입 뷰 이동 및 수정
